chore: move wallet to context
