+++ src/context/wallet.tsx
@@ -1,9 +1,9 @@
 import React, { useContext, useEffect, useMemo, useState } from "react";
 import Wallet from "@project-serum/sol-wallet-adapter";
-import { notify } from "./notifications";
-import { useConnectionConfig } from "./connection";
-import { useLocalStorageState } from "./utils";
-import { SolongAdapter } from "./solong_adapter";
+import { notify } from "../utils/notifications";
+import { useConnectionConfig } from "../utils/connection";
+import { useLocalStorageState } from "../utils/utils";
+import { SolongAdapter } from "../utils/solong_adapter";
 
 export const WALLET_PROVIDERS = [
   { name: "sollet.io", url: "https://www.sollet.io" },
