+++ src/utils/accounts.tsx
@@ -6,7 +6,7 @@ import React, {
   useState,
 } from "react";
 import { useConnection } from "./connection";
-import { useWallet } from "./wallet";
+import { useWallet } from "../context/wallet";
 import { AccountInfo, Connection, PublicKey } from "@solana/web3.js";
 import { programIds, SWAP_HOST_FEE_ADDRESS, WRAPPED_SOL_MINT } from "./ids";
 import { AccountLayout, u64, MintInfo, MintLayout } from "@solana/spl-token";
